1. CommunityServiceImpl.fixGarbledUsers() 引用了不存在的 user 表
  CommunityServiceImpl.java:135-139 — SQL 语句是 UPDATE user SET name = ...，但实际的表名是 sys_user。user 表在 V2 升级脚本(database_upgrade_v2.sql:49) 中被删除了：DROP TABLE IF EXISTS
  user;。运行此修复方法将抛出 SQLException。

  // 错误 (第135行):
  jdbcTemplate.update("UPDATE user SET name = '管理员', ...");
  // 应为:
  jdbcTemplate.update("UPDATE sys_user SET name = '管理员', ...");

  2. HealthData 实体对 DECIMAL 列使用了 Double

  HealthData.java:29-39 — V2 升级将这些列转换为 DECIMAL 以修复精度问题，但 Java 实体仍然使用 Double（浮点数）：

  ┌─────────────────┬───────────┬──────────────┬────────────┐
  │      字段       │ Java 类型 │   DB 类型    │    应为    │
  ├─────────────────┼───────────┼──────────────┼────────────┤
  │ bloodSugar      │ Double    │ decimal(5,2) │ BigDecimal │
  ├─────────────────┼───────────┼──────────────┼────────────┤
  │ bodyTemperature │ Double    │ decimal(4,2) │ BigDecimal │
  ├─────────────────┼───────────┼──────────────┼────────────┤
  │ calories        │ Double    │ decimal(8,2) │ BigDecimal │
  └─────────────────┴───────────┴──────────────┴────────────┘

  这与 CommunityService.price 和 ServiceOrder.amount 已经做的 BigDecimal 修复不一致。

  ---
  高优先级问题

  3. 活动注册中的乐观锁竞态条件

  ActivityServiceImpl.java:113-114 — registerActivity() 方法在调用 activityMapper.updateById(activity) 后未检查返回值。使用 @Version 乐观锁时，MyBatis-Plus 在版本冲突时返回
  0。如果两个并发注册发生冲突：
  - 一个更新静默失败（0 行）
  - 但参与者记录已经插入
  - 结果：current_participants 对实际参与者计数不足

  将此与 WalletServiceImpl.recharge() 对比，后者正确检查了 rows == 0。

  4. PaymentController.getElderOrders() 仅获取第一个老人的订单

  PaymentController.java:151-152:
  List<Long> elderIds = relations.stream().map(UserRelation::getElderId).collect(Collectors.toList());
  Map<String, Object> orderQueryMap = new HashMap<>();
  orderQueryMap.put("userId", elderIds.get(0));  // BUG: 仅获取第一个！

  这会静默地遗漏所有其他老人的订单。应使用 IN 子句或遍历所有老人 ID。

  5. DatabaseInitializer 和 DatabaseDataFixer 完全重复

  两个类都实现了 CommandLineRunner 并包含几乎相同的数据修复方法。它们在启动时同时运行，使每个修复操作执行两次。这是冗余的，如果它们交错执行，可能会在 fillWalletTable
  中引起竞态条件（两者都可能尝试为同一用户插入钱包行）。

  ---
  中优先级问题

  6. PaymentController 直接注入 Mapper 层

  PaymentController.java:25,31 — ServiceOrderMapper 和 UserRelationMapper
  被直接注入到控制器中，绕过了服务层。项目中的每个其他控制器都通过服务层操作。这违反了分层架构，使得事务管理、测试和关注点分离更加困难。

  7. ServiceOrderServiceImpl.updateOrderStatus() 和 cancelOrder() 忽略乐观锁结果

  ServiceOrderServiceImpl.java:64-71, 74-81 — 两个方法都调用 updateById(order) 但未检查返回值。ServiceOrder 实体有 @Version，因此在并发更新下，一个调用可能静默失败。

  8. DeviceServiceImpl.controlDevice() 在控制设备时不更新 lastOnlineTime

  DeviceServiceImpl.java:82-90 — 与设置 lastOnlineTime 的 updateDeviceStatus() 方法不同，controlDevice() 遗漏了它。这是个语义问题 — 当设备被控制时，它肯定是在线的。

  ---
  低优先级问题

  9.  fixGarbledUsers 硬编码了错误的用户 ID

  CommunityServiceImpl.java:134-140 — 行如 UPDATE user SET name = '张大爷', username = 'elder1' ... WHERE id = 2 硬编码了与当前数据库数据不匹配的名称和用户名（例如，ID 2 实际上是 testuser /
  张大爷，不是 elder1）。

  10. EmergencyCallMapper.findUnresolvedCalls() 包含了已取消的呼叫

  查询 WHERE status != 'resolved' 返回状态为 pending、responding 和 cancelled 的呼叫 — 已取消的呼叫可能不应出现在"未解决"列表中。应为：WHERE status IN ('pending', 'responding')。

  11. DatabaseInitializer.insertActivities() 为新活动设置了非零的 current_participants

  DatabaseInitializer.java:351 — int[] currentParticipants = {15, 42, 28, 66, 31} 用于种子数据插入，但新活动应从 0 名参与者开始。这与实际参与者记录不匹配。

  12. WalletServiceImpl 使用 @Autowired 而其余地方使用 @RequiredArgsConstructor

  不一致的注入风格 — 项目中所有其他服务都使用 Lombok 的 @RequiredArgsConstructor 配合 final 字段。

  13. User 实体中的 autoResultMap = true 是不必要的

+++ backend/src/main/java/com/example/smartcommunity/entity/HealthData.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -14,35 +15,38 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @TableName("health_data")
 public class HealthData {
-    
+
     @TableId(type = IdType.AUTO)
     private Long id;
-    
+
     private Long userId;
-    
+
     private Integer heartRate;
-    
+
     private Double bloodPressureHigh;
-    
+
     private Double bloodPressureLow;
-    
-    private Double bloodSugar;
-    
-    private Double bodyTemperature;
-    
+
+    // 修复精度丢失：与数据库 decimal(5,2) 匹配
+    private BigDecimal bloodSugar;
+
+    // 修复精度丢失：与数据库 decimal(4,2) 匹配
+    private BigDecimal bodyTemperature;
+
     private Integer sleepHours;
-    
+
     private Integer sleepQuality;
-    
+
     private Integer steps;
-    
-    private Double calories;
-    
+
+    // 修复精度丢失：与数据库 decimal(8,2) 匹配
+    private BigDecimal calories;
+
     private String healthReport;
-    
+
     private Integer warningLevel;
-    
+
     private LocalDateTime measuredAt;
-    
+
     private LocalDateTime createdAt;
 }
